+++ game/src/lib.rs
@@ -1,3 +1,12 @@
+
+// This file is part of EXA.
+// EXA is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as 
+// published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+// EXA is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty
+// of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License along with EXA. If not, see 
+// <https://www.gnu.org/licenses/>.
+
 //!
 //! Game project.
 //! 
@@ -8,7 +17,7 @@
 pub mod player;     /// Player object and script.
 pub mod settings;   /// Player/game settings.
 pub mod ui;         /// Game User Interface.
-mod tracy;          /// Tracy utilities and subroutines.
+// pub mod eventline;  /// Events processor subroutine.
 mod utilities;      /// Game utilities.
 
 #[cfg(feature = "tracy")]
@@ -20,14 +29,9 @@ use fyrox::{
         pool::Handle,
         reflect::prelude::*,
         visitor::prelude::*
-    },
-    event::Event,
-    gui::{ message:: { MessageDirection, UiMessage }, text::{Text, TextMessage}, UiNode },
-    plugin::{ Plugin, PluginContext, PluginRegistrationContext },
-    scene::Scene
+    }, event::{DeviceEvent, Event}, gui::{ message:: { MessageDirection, UiMessage }, text::{Text, TextMessage}, UiNode }, keyboard:: { PhysicalKey, KeyCode }, plugin::{ Plugin, PluginContext, PluginRegistrationContext }, scene::Scene
 };
-use std::path::Path;
-use { tracy_client, tracy_client_sys };
+use std:: { future::{Future, IntoFuture}, path::Path };
 use tracing:: { trace, trace_span, debug, debug_span, info, info_span, warn, warn_span, error, error_span, instrument };
 use crate::utilities::*;
 
@@ -38,6 +42,8 @@ use { tracy_client, tracy_client_sys };
 const GAME_TITLE:       &'static str    = "Experiment A";
 /// Game version is read in package file `Cargo.toml`.
 const GAME_VERSION:     &'static str    = env!("CARGO_PKG_VERSION");
+/// Save file path.
+const SAVE_FILEPATH:    &'static str    = "./save0.bin";
 
 
 #[derive(Debug, Reflect, Visit, Default)]
@@ -49,23 +55,76 @@ pub struct Game {
     /// User Interfaces.
     ui:             ui::UiSubset,
 
+    /// Game Settings and the Settings UI.
+    #[visit(skip)]
+    #[reflect(hidden)]
+    settings:       settings::Settings,
+
+    #[visit(skip)]
+    #[reflect(hidden)]
+    gamepads:       Option<gilrs::Gilrs>,
+
 }
 
 impl Game {
 
-    #[instrument(skip(context))]
-    pub fn new(scene_path: Option<&str>, mut context: PluginContext) -> Self {
+    // #[instrument(skip(context))]
+    // pub fn new(scene_path: Option<&str>, mut context: PluginContext) -> Self {
 
-        trace!("Creating game plugin context...");
+    //     trace!("Creating game plugin context...");
 
-        context
-            .async_scene_loader
-            .request(scene_path.unwrap_or("data/scene.rgs"));
+    //     context
+    //         .async_scene_loader
+    //         .request(scene_path.unwrap_or("data/scene.rgs"));
 
-        Self {
-            scene:      Handle::NONE,
-            ui:         ui::UiSubset::new(&mut context)
-        }
+    //     Self {
+    //         scene:      Handle::NONE,
+    //         ui:         ui::UiSubset::new(&mut context),
+    //         settings:   settings::Settings::default()
+    //     }
+    // }
+
+    fn save(&self, context: &mut PluginContext<'_, '_>) -> VisitResult {
+
+        let mut visitor = Visitor::new();
+        context.scenes[self.scene].save("Scene", &mut visitor)?;
+        // TODO: Add game fields as necessary.
+        visitor.save_binary(SAVE_FILEPATH)
+
+    }
+
+    fn load(&mut self, context: &mut PluginContext<'_, '_>) {
+        context.async_scene_loader.request_raw(SAVE_FILEPATH);
+    }
+
+    // TODO: Document.
+    fn on_suspended(&self) {
+
+        // TODO: Pause game.
+
+    }
+
+    // TODO: Document.
+    fn on_resumed(&self) {
+
+        // TODO: Continue game.
+
+    }
+
+    // TODO: Document.
+    fn on_loopexiting(&self) {
+
+        // TODO: Deinit things and run save routines where necessary.
+
+    }
+
+    // TODO: Document.
+    fn on_memory_warning(&self) {
+        
+        // TODO: Consider lessening the memory footprint wherever possible.
+        // ? This doesn't do much besides warn the application user that the app may be terminated soon.
+        error!("Memory Warning: OS indicates memory usage exceeds limits and may terminate this game soon!");
+
     }
 
 }
@@ -79,12 +138,21 @@ impl Plugin for Game {
 
     #[instrument(
         name = "Plugin Update",
-        skip(context)
+        skip(context, self)
     )]
     fn update(&mut self, context: &mut PluginContext) {
 
-        // Retrieve Tracy client handle.
-        let _tracy = tracy_client::Client::running();
+        // Run gamepad input updates.
+        if let Some(gilrs) = &mut self.gamepads {
+
+            while let Some(event) = gilrs.next_event() {
+                trace!("gilrs event by {id}: {ev:?}", id = event.id, ev = event.event);
+            };
+
+            // Increment event counter.
+            // * This must be called after all event processing with **Gilrs**!
+            gilrs.inc();
+        }
 
         // Run UI updates.
         self.ui.update(context);
@@ -94,12 +162,12 @@ impl Plugin for Game {
 
             #[cfg(feature="tracy")]
             {
-            // If Tracy is running, collect a frame image.
+                // If Tracy is running, collect a frame image.
                 let _tracy = tracy_client::Client::running();
-            if _tracy.is_some() {
-                let _tracy = _tracy.unwrap();
-                tracy::frameimage_collect(igc);
-                _tracy.frame_mark();
+                if _tracy.is_some() {
+                    let _tracy = _tracy.unwrap();
+                    tracy::frameimage_collect(igc);
+                    _tracy.frame_mark();
                 }
             }
 
@@ -111,7 +179,7 @@ impl Plugin for Game {
     fn on_os_event(
         &mut self,
         _event: &Event<()>,
-        _context: PluginContext,
+        mut _context: PluginContext,
     ) {
 
         match _event {
@@ -120,27 +188,49 @@ impl Plugin for Game {
                 // For reading cause of incoming events.
             },
 
-            // Event::WindowEvent { 
-            //     window_id, event
-            // }                       => {
-            //     // Window events.
-            // },
+            Event::WindowEvent { 
+                window_id, event
+            }                       => {
 
-            // Event::DeviceEvent {
-            //     device_id, event
-            // }                       => todo!(),
+            },
+
+            // TODO: Remove after testing load/save functionality.
+            Event::DeviceEvent { device_id, event } => {
+                if let DeviceEvent::Key(key) = event {
+                    
+                    // Performs a save/load on F2/F3 key press.
+                    if let PhysicalKey::Code(kc) = key.physical_key {
+                        match kc {
+                            
+                            KeyCode::F2     => {
+                                info!("Saving game; F2 key...");
+                                self.save(&mut _context);
+                            },
+
+                            KeyCode::F3     => {
+                                info!("Loading game; F3 key...");
+                                self.load(&mut _context);
+                            },
+
+                            _               => {}
+
+                        }
+                    }
+
+                }
+            },
 
             // Event::UserEvent(_)     => todo!(),
 
-            // Event::Suspended        => todo!(),
+            Event::Suspended        => self.on_suspended(),
 
-            // Event::Resumed          => todo!(),
+            Event::Resumed          => self.on_resumed(),
 
             // Event::AboutToWait      => todo!(),
 
-            // Event::LoopExiting      => todo!(),
+            Event::LoopExiting      => self.on_loopexiting(),
 
-            // Event::MemoryWarning    => todo!(),
+            Event::MemoryWarning    => self.on_memory_warning(),
 
             _ => {} // Ignore other event types.
 
@@ -185,6 +275,9 @@ impl Plugin for Game {
         // Report and set the scene into `self`.
         info!("Scene ({scene:?}) loaded: {path}", scene = new_scene, path = path.display());
         self.scene = new_scene;
+        if let Ok(mut visitor)= Visitor::load_from_memory(data) {
+            // TODO: Take `data` and visit `self` fields.
+        }
 
     }
     
@@ -192,11 +285,8 @@ impl Plugin for Game {
     fn register(&self, context: PluginRegistrationContext) {
 
         // TODO: Register scripts here.
-
-        // context
-        //     .serialization_context
-        //     .script_constructors
-        //     .add::<Sequencer>("Sequencer");
+        let script = &context.serialization_context.script_constructors;
+        script.add::<player::Player>("Player");
 
     }
     
@@ -206,7 +296,7 @@ impl Plugin for Game {
         fyrox::gui::inspector::editors::PropertyEditorDefinitionContainer::empty()
     }
     
-    #[instrument(skip(context))]
+    #[instrument(skip(context, self))]
     fn init(
         &mut self,
         scene_path: Option<&str>,
@@ -220,6 +310,18 @@ impl Plugin for Game {
             .request(scene_path.unwrap_or("data/scene.rgs"));
 
         self.ui = ui::UiSubset::new(&mut context);
+
+        // Attempt to retrieve gamepads from the system.
+        if let Ok(gilrs) = gilrs::Gilrs::new() {
+            
+            // ? Indicate what gamepads are connected.
+            for (each_gamepad_id, _each_gamepad) in gilrs.gamepads() {
+                info!("connected gamepad: {}", each_gamepad_id);
+            }
+
+            self.gamepads = Some(gilrs);
+
+        }
 
     }
     
@@ -235,7 +337,13 @@ impl Plugin for Game {
     ) {
 
         trace!("Graphics context initialized!");
-        // TODO: Register tracy framecollector render pass.
+        // // TODO: Register tracy framecollector render pass.
+        // utilities::with_igc(context.graphics_context, |igc| {
+            
+        //     let rc_tracy_framecollecter = std::rc::Rc::new(std::cell::RefCell::new(tracy::FrameCollector::default()));
+
+        //     igc.renderer.add_render_pass(rc_tracy_framecollecter);
+        // });
 
     }