+++ CHANGES.md
@@ -21,6 +21,15 @@ This is the inception history of the project. Pretty bare-bones around here...
 - Introduces dependencies:
     - [strum](https://crates.io/crates/strum) v0.26.3
     - [strum_macros](https://crates.io/crates/strum_macros) v0.26.3
+    - [gilrs](https://crates.io/crates/gilrs) v0.11.0
+    - [serde](https://crates.io/crates/serde) v1.0.210
+    - [ron](https://crates.io/crates/ron) v0.8.1
+    - [fyrox](https://crates.io/crates/fyrox) v0.34.0
+    - [firedbg-lib](https://crates.io/crates/firedbg-lib) v0.1.2
+    - [tracing](https://crates.io/crates/tracing) v0.1.40
+    - [tracy-client](https://crates.io/crates/tracy-client) v0.17
+    - [tracy-client-sys](https://crates.io/crates/tracy-client-sys) v0.22
+    - [fluent-bundle](https://crates.io/crates/fluent-bundle) v0.15.3
 
 - Start of CI/CD on [GitHub repo](https://github.com/lilyanavalley/exa) with files:
     - [.github/workflows/test.yaml](./.github/workflows/test.yaml) for testing game on push or P.R.